+++ src/main/kotlin/br/com/storehouse/data/model/VendaRequest.kt
@@ -1,5 +1,7 @@
 package br.com.storehouse.data.model
 
+import java.math.BigDecimal
+
 data class VendaRequest(
     val itens: List<ItemVendaRequest>,
     val pagamentos: List<PagamentoVendaRequest>
@@ -12,5 +14,5 @@ data class ItemVendaRequest(
 
 data class PagamentoVendaRequest(
     val tipo: String,
-    val valor: Double
+    val valor: BigDecimal
 )
